backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Pavlov Danil",email = "danpavkzn@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite (==0.21.0)",
    "alembic (==1.15.1)",
    "annotated-types (==0.7.0)",
    "anyio (==4.8.0)",
    "black (==25.1.0)",
    "certifi (==2025.1.31)",
    "click (==8.1.8)",
    "colorama (==0.4.6)",
    "dnspython (==2.7.0)",
    "email-validator (==2.2.0)",
    "fastapi (==0.115.8)",
    "fastapi-cli (==0.0.7)",
    "greenlet (==3.1.1)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "jinja2 (==3.1.5)",
    "mako (==1.3.9)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==3.0.2)",
    "mdurl (==0.1.2)",
    "mypy-extensions (==1.0.0)",
    "packaging (==24.2)",
    "pathspec (==0.12.1)",
    "platformdirs (==4.3.6)",
    "pydantic (==2.10.6)",
    "pydantic-settings (==2.8.1)",
    "pydantic-core (==2.27.2)",
    "pygments (==2.19.1)",
    "python-dotenv (==1.0.1)",
    "python-multipart (==0.0.20)",
    "pytz (==2025.1)",
    "pyyaml (==6.0.2)",
    "rich (==13.9.4)",
    "rich-toolkit (==0.13.2)",
    "shellingham (==1.5.4)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.38)",
    "sqlmodel (==0.0.22)",
    "starlette (==0.45.3)",
    "typer (==0.15.1)",
    "typing-extensions (==4.12.2)",
    "uvicorn (==0.34.0)",
    "watchfiles (==1.0.4)",
    "websockets (==15.0)",
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
